Fix tests originally ported from papertrail module
